+++ UpgradeWorld/commands/LocationsAdd.cs
@@ -12,7 +12,7 @@ public class LocationsAddCommand
       LocationIdParameters pars = new(args);
       if (Helper.IsClient(args)) return;
       if (!pars.Valid(args.Context)) return;
-      Executor.AddOperation(new DistributeLocations(args.Context, pars.Ids, pars.Start, pars.Chance));
+      Executor.AddOperation(new DistributeLocations(args.Context, pars.Ids, pars));
       Executor.AddOperation(new SpawnLocations(args.Context, pars.Ids, pars));
     }, optionsFetcher: () => ZoneSystem.instance.m_locations.Select(location => location.m_prefabName).ToList());
   }