+++ UpgradeWorld/filterers/LocationFilterer.cs
@@ -7,7 +7,6 @@ public class LocationFilterer(IEnumerable<string> Ids, bool OnlyMissing) : IZone
   public Vector2i[] FilterZones(Vector2i[] zones, ref List<string> messages)
   {
     var locationObjects = Ids.Select(id => id.GetStableHashCode()).ToHashSet();
-    if (locationObjects.Count == 0) return zones;
     var zs = ZoneSystem.instance;
     var amount = zones.Length;
     zones = zones.Where(zone =>