+++ publish/CHANGELOG.md
@@ -1,7 +1,9 @@
 - v1.41
   - Changes `world_clean` to optimize dungeons instead of removing extra rooms.
+  - Changes the location distribution to be affected by safe zones.
   - Fixes `locations_add` command placing every unplaced location, instead of just the specified locations (mostly affects some custom maps).
   - Fixes `world_clean` command not working for new chest format.
+  - Fixes `locations_remove` command not being affected by safe zones.
   - Improves the performance `locations_add` command (used by most upgrade commands).
 
 - v1.40